# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions": {
      "post": {
        "operationId": "SessionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session name already exists"
          }
        },
        "summary": "Create a new WhatsApp session",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "operationId": "SessionController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max sessions to return (1-1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of sessions to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "operationId": "SessionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get session by ID",
        "tags": [
          "sessions"
        ]
      },
      "delete": {
        "operationId": "SessionController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/start": {
      "post": {
        "operationId": "SessionController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session already started"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Start a session and initialize WhatsApp connection",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/stop": {
      "post": {
        "operationId": "SessionController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session stopped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Stop a session and disconnect WhatsApp",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/force-kill": {
      "post": {
        "operationId": "SessionController_forceKill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session force-killed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Force-kill a stuck session (SIGKILL its wedged engine, then tear it down)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/proxy": {
      "patch": {
        "description": "Persist the per-session egress proxy from structured fields (send no host to clear it). Takes effect on the next start; restart the session to apply immediately. For a credentialed SOCKS5 proxy, OpenWA runs an in-process authenticating relay so the Chromium engine can egress through it.",
        "operationId": "SessionController_updateProxy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProxyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proxy updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Set or clear a session proxy",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/proxy/verify": {
      "post": {
        "description": "Fetch the public IP through the session's configured proxy and compare it against the box's direct IP, confirming traffic actually leaves through the tunnel.",
        "operationId": "SessionController_verifyProxy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification result"
          },
          "400": {
            "description": "No proxy configured for this session"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Verify a session egresses through its proxy",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/qr": {
      "get": {
        "operationId": "SessionController_getQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "QR code not ready or session already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get QR code for session authentication",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/pairing-code": {
      "post": {
        "operationId": "SessionController_requestPairingCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing code generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not started or already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Request an 8-char pairing code to link via phone number (alternative to QR)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/groups": {
      "get": {
        "operationId": "SessionController_getGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max groups to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of groups to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups the session is a member of"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all groups for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats": {
      "get": {
        "operationId": "SessionController_getChats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max chats to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active chats (most recent first)"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get active chats for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/read": {
      "post": {
        "operationId": "SessionController_markChatRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as read successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as read/seen",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/unread": {
      "post": {
        "operationId": "SessionController_markChatUnread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as unread successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as unread",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/delete": {
      "post": {
        "operationId": "SessionController_deleteChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat deleted successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a chat from the chat list (e.g. a group you have left)",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/groups": {
      "post": {
        "operationId": "GroupController_create",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        },
        "summary": "Create a new group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/stats/sessions/{sessionId}": {
      "get": {
        "operationId": "StatsController_getSessionStats",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-session statistics for the requested session."
          }
        },
        "summary": "Get statistics for a specific session",
        "tags": [
          "statistics"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.10.6",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name for the session (alphanumeric and hyphens only)",
            "example": "my-bot",
            "minLength": 3,
            "maxLength": 50
          },
          "config": {
            "type": "object",
            "description": "Session configuration options. Set autoRejectCalls (boolean, default false) to automatically reject incoming calls — the call.received event is still emitted.",
            "example": {
              "autoReconnect": true
            }
          },
          "proxyUrl": {
            "type": "string",
            "description": "Optional per-session egress proxy URL (http/https/socks4/socks5; credentialed form \"http://user:pass@host\" allowed). Must be a REAL, REACHABLE proxy — an unreachable value silently blocks the WhatsApp WebSocket (no QR is ever delivered) and the session start times out (~30s → 504 Gateway Timeout). Leave unset unless your network cannot reach WhatsApp directly."
          },
          "proxyType": {
            "type": "string",
            "description": "Proxy type",
            "enum": [
              "http",
              "https",
              "socks4",
              "socks5"
            ],
            "example": "http"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "QRCodeResponseDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "QR code as data URL",
            "example": "data:image/png;base64,..."
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "initializing",
              "qr_ready",
              "authenticating",
              "ready",
              "disconnected",
              "failed"
            ],
            "example": "qr_ready"
          }
        },
        "required": [
          "qrCode",
          "status"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}